Rework options -m, -n to be min and max
The previous overloading of -n as a number, range, or flag to read from stdin
caused several problems.  First, it was complicated.  More importantly though
it allowed for arbitrary values to come in from stdin.  That implicitly forces
series to behave like functions, which is bad for series like fibonacci.

This ensures series behave like series.  If the user wants to limit the number
output, then they can use head.